feat: optimize the database persistence and loading and implements the telemetry to the watcher module
main

+++ core/src/domain/dtos/adherence_test.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "camelCase")]
 pub struct AdherenceTest {
     /// The unique identifier for the clade to be tested.
-    pub clade: UntaggedParent<Clade, i32>,
+    pub clade: UntaggedParent<Clade, u64>,
 
     /// The number of kmer matches with the desired clade.
     pub one_len: i32,
